feat: Добавлены сущности: Вид измерений (MeasCategory), Характаристика СИ (MiCharacteristic), Состояние СИ (MiCondition), Статус СИ (MiStatus), Подразделение (Department) а также  контроллеры сервисы и пр. классы необходимые для взаимодействия с данными сущностями. Свойства класса MeasurementInstrument распределены между классами MeasurementInstrument и MiDetails. Для работы с MiDetails добавлен MiDetailsService. Для преобразования сущностей в DTO добавлена новая зависимость ModelMapper.

+++ src/main/java/main/controller/MiStandardController.java
@@ -87,7 +87,7 @@ public class MiStandardController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteMiStandard(@PathVariable("id") int id){
+    public ResponseEntity<?> deleteMiStandard(@PathVariable("id") long id){
         miStandardService.delete(id);
         String okMessage ="Запись об эталоне № " + id + " успешно удалена";
         log.info(okMessage);
